requirements.txt
pyyaml>=6.0.0
click>=8.1.7
pydantic>=2.0.0
jinja2>=3.0.0 